refactor(gradle): extract registry content parsing
* to make it reusable
