+++ lib/modules/manager/gradle/parser/registry/content.spec.ts
@@ -1,0 +1,6 @@
+describe('modules/manager/gradle/parser/registry/content', () => {
+  // TODO: Add tests
+  it('todo', () => {
+    expect(1).toBe(1);
+  });
+});
